Set up the test for the create many applied roles
